# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AccountResponse" : {
        "properties" : {
          "balance" : {
            "type" : "string"
          },
          "iban" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "FROZEN" ],
            "type" : "string"
          },
          "type" : {
            "enum" : [ "CHECKING", "SAVINGS", "LOAN" ],
            "type" : "string"
          }
        },
        "required" : [ "balance", "iban", "id", "status", "type" ],
        "type" : "object"
      },
...
      "CardResponse" : {
        "properties" : {
          "expMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "expYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "last4" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "FROZEN" ],
            "type" : "string"
          }
        },
        "required" : [ "expMonth", "expYear", "id", "last4", "status" ],
        "type" : "object"
      },
...
      "IssuedCardResponse" : {
        "properties" : {
          "cvv" : {
            "type" : "string"
          },
          "expMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "expYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "pan" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "FROZEN" ],
            "type" : "string"
          }
        },
        "required" : [ "cvv", "expMonth", "expYear", "id", "pan", "status" ],
        "type" : "object"
      },
...
      "TransactionResponse" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "fromIban" : {
            "description" : "null on deposits/credits (no originator)",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "TRANSFER", "DEPOSIT", "INTEREST" ],
            "type" : "string"
          },
          "memo" : {
            "type" : [ "string", "null" ]
          },
          "postedAt" : {
            "description" : "null while HELD or when CANCELLED - money has not moved",
            "type" : [ "string", "null" ]
          },
          "reviewed" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "POSTED", "HELD", "CANCELLED" ],
            "type" : "string"
          },
          "toIban" : {
            "description" : "null on engine loan charges (no destination)",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "createdAt", "currency", "flagged", "fromIban", "id", "kind", "memo", "postedAt", "reviewed", "status", "toIban" ],
        "type" : "object"
      },
      "TransferRequest" : {
        "properties" : {
          "amount" : {
            "pattern" : "^\\d+(\\.\\d{1,4})?$",
            "type" : "string"
          },
          "currency" : {
            "pattern" : "(?i)\\s*USD\\s*",
            "type" : "string"
          },
          "fromAccountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memo" : {
            "maxLength" : 140,
            "minLength" : 0,
            "type" : "string"
          },
          "toIban" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "amount", "toIban" ],
        "type" : "object"
      },
      "TransferResponse" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "fromIban" : {
            "description" : "null on deposits (no originator)",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memo" : {
            "type" : [ "string", "null" ]
          },
          "postedAt" : {
            "description" : "null while HELD or when CANCELLED - money has not moved",
            "type" : [ "string", "null" ]
          },
          "status" : {
            "enum" : [ "POSTED", "HELD", "CANCELLED" ],
            "type" : "string"
          },
          "toIban" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "createdAt", "currency", "flagged", "fromIban", "id", "memo", "postedAt", "status", "toIban" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Full-stack online banking demo. Errors are RFC-7807 (ApiProblem); money travels as JSON strings.",
    "title" : "Online Banking Simulator API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/accounts/{id}/cards" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CardResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "card-controller" ]
      },
      "post" : {
        "operationId" : "issue",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedCardResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "card-controller" ]
      }
    },
...
    "/api/v1/admin/accounts/{id}/unfreeze" : {
      "post" : {
        "operationId" : "unfreeze_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      }
    },
...
    "/api/v1/cards/{id}/freeze" : {
      "post" : {
        "operationId" : "freeze",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "card-controller" ]
      }
    },
    "/api/v1/cards/{id}/unfreeze" : {
      "post" : {
        "operationId" : "unfreeze",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "card-controller" ]
      }
    },
...
    "/api/v1/operations" : {
      "get" : {
        "operationId" : "operation",
        "parameters" : [ {
          "in" : "query",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
...
    "/api/v1/transfers" : {
      "post" : {
        "operationId" : "transfer",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransferRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransferResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ]
}